requirements.txt
flask
requests
google-cloud-storage
dominate